CXXCBC-54 Add log forwarding
Motivation
==========
When the client is used in the sdk wrappers (python, node, etc...), it
would be nice to allow the logging to be forwarded to the wrapper's
logging handlers.

Modification
============
Simplest thing to do is add a spdlog::sinks::sink to the logger
configuration.  If it is set, we add it.   Then the wrappers can
implement a sink that does the handoff.

Results
=======
I didn't add tests, as there are no logging tests anyways.  I did for
the transactions logging, and I've used this in python.