CXXCBC-55: External Tracing and Metrics support (#228)
* CXXCBC-55: External Tracing and Metrics support

Motivation
==========
Allow clients to plug in their own tracing and/or metrics solutions,
such as OpenTelemetry

Modifications
=============
Expose the tracer and meter in cluster_options.  Then, add a parent_span
to many (but for now, not all) requests.   The tests use a simple tracer
and meter implementation to facilitate testing.

*** Will update this with a header that wraps OpenTelemetry objects in
particular, which will not be referenced by anything in the client so we
will only require otel when it is used.

Results
=======
Tests pass

* fixup tls in integration_test_guard

* Second pass at correcting tests

* fixup tracer integration test