CXXCBC-55 - OpenTelemetry support
Motivation
==========
Add support to get tracing and metrics data into OpenTelemetry
providers.  Since we don't depend on otel libs or headers, these
are not referenced by any other code in the client.

Modification
============
Added headers to wrap an OpenTelemetry Tracer and Span, and one to wrap
a Meter and Histogram.  Created a stand-alone example to use each, which
will be incorporated into the source tree later when we have an examples
directory.

Results
=======
Used in my otel examples successfully.