TXNCXX-265: ExtSDKIntegration
Motivation
==========
Now that transactions is integrated into the sdk, we needed to update it
to support the ExtSDKIntegration changes.

Modification
============
Configs have changed names, and the cleanup no longer opens every
bucket, and assumes the atrs are in the default collection.    Each time
we create an atr, we update the cleanup to clean that keyspace, if it is
a new one.

Also - changed a json file, and noticed our .clang-format doesn't like
json, so added a section for that.