CXXCBC-251: fix snappy decompression for get_replica
The library incorrectly calculates offset of the body when alternative
response layout is being used (when flexible extra frames presented).

For regular get operations it is not visible, but get_replica responses
contain key, and wrong offset cause snappy uncompress function to give
up, and the application receives compressed payload. Which cannot be
parsed as JSON.