CXXCBC-257 Support for FIT testing of kv range scan
Since FIT uses the public api, we need (at least for now) a way to get
the core cluster object from the public cluster object, to use kv range
scan.   Added an accessor for that, for now.

Also added a check to apply the limit on a sampling scan, so next()
stops when the limit is reached.

And a small logging change so we can see the collection being refreshed.