Eliminate looping transform in mcbp_parser::next
I found that in mcbp_session_impl::do_read, 39% of the time was in
mcbp::parser::next, with the vast majority of that in the std::transform
call after uncompressing the data.   By using vector::insert instead,
that dropped to only 11% of the time in do:read.

Probably the upshot here is to look for everywhere we are filling
containers with a back_inserter and considering a way to modify that to
not loop and instead copy in blocks.