CXXCBC-310 Transaction fit test errors
When we stage an insert, and get an ambiguous fail back, we look for the
doc and examine it to see if we can retry, etc...   When doing so, if
the doc was staged successfully, we just need to make sure the that we
use the staged_mutation's content, not the content in the
staged_mutation's doc, as that will have the results of looking at the
document after we staged it as a tombstone, so the content will be
empty.   This showed up in some ambiguous fail tests in fit.  Simple
fix.