CXXCBC-310 LostTxnCleanup wasn't stopping right away
When shutting down, for instance, the lost attempt loop would not stop
sometimes, leading to issues with tests sometimes.  Now, I'm only
accessing the running_ bool under the lock, so it is safe to use it in
the wait calls as well.  Seems to have been an issue, simple fix.

Also - I was creating a client record doc if it was missing, when
deleting.   Seems unnecessary.