CXXCBC-320 Negative expiry in atr can 'stick' docs. (#393)
The effect is that the expiry seems to be a very large positive number.
Lets just read it as an int32_t, cap it at 0 on the low side.  Turns out
we definitely cap at 0 when we write, so this probably has to be from
some old version that had flawed logic when writing.