CXXCBC-387: Performance: potential micro-optimisation on tags and local/remote sessions
Profiling indicates that non-trivial time is spent assigning
tags to spans, which is wasted effort if noop_tracer is being
used.

Borrowing an optimisation from the JVM SDKs, where we skip this
assignment in this case.

This optimisation, together with another small one that will be
submitted separately, produced around a 15% performance improvement.