CXXCBC-403: Allow retries for KV not_my_vbucket response
Motivation
----------
Currently if a KV operation receives a NMVB, it is possible the
operation will raise a request_canceled error. When the client receives
a NMVB there are circumstances in which we should attempt to retry the
operation instead of raising a request_canceled.

Changes
-------
Update logic for handling NMVB KV responses so that retries are possible
when appropriate.