CXXCBC-368: Subscribe to clustermap notifications to speedup failover (#490)
* cbc-beam tool to send operations to specified vbucket/node

* CXXCBC-368: Subscribe to clustermap notifications to speedup failover

Implement only deduplication flag from the RFC, and enable clustermap
change notifications. I this patch, we only want to use full
notifications, and do not turn off polling. So the SDK will be able to
receive new configuration sooner, but still be chatty as older versions.

* detect sequence change in the node lists and fix the order

It is possible that we will receive configuration, where nodes neither
removed, nor added, but still we need to check the order of them in
nodeExt, because at early stages of rebalance, ns_server might add new
nodes to the end of the nodesExt list (while not adding them to
serverList at all). In this case we don't want to keep this temporary
order forever, and should be able to detect this condition an reorder
them later.