CXXCBC-461: Do not send ping to nodes that have not completed bootstrap. (#540)
If doing a ping during a rebalance of a multi-node cluster it is
possible that one or more of the nodes are in the process of completing
bootstrap.  This can cause the ping operation to timeout.  Since we know
when a node has bootstrapped or not, the SDK should not ping a node that has
not bootstrapped.  This will allow for the ping operation to complete
without the worry of a timeout due to waiting on the bootstrap process to successfully complete.

* Add a mechanism to keep track of the last error w/in the
  `bootstrap_handler`.
* If ping is called, check if the node has bootstrapped.  If the node is
  not bootstrapped, add the kv node to the ping report with an error
indicating either the last error message from the `bootstrap_handler` or
a generic message, but do not execute the ping operation.