CXXCBC-448: Expose hooks for fork() scenarious (#536)
* CXXCBC-448: Expose hooks for fork() scenarious

In order to gracefully survive fork() syscall, the library should be
able to move itself into stable and predictable state. In particular
ASIO should not have any pending work, and all IO thread should be
stopped. After the fork() the parent process retains control over all
file descriptors, but the child have to reconnect them (ASIO does it in
its hooks), and also the SDK should restart all the IO workers.

At the moment, the only worker thread that the library controls is the
one that in charge of the transactions cleanup, but later once the
public API will hide ASIO context, it will become in charge of the
restarting IO threads.

In case of the wrappers, that use core API directly, they must be able
to stop all IO threads before fork(), wait for work completion, then
perform the syscall and restart IO threads and contexts in both child
and parent later. See example in test/test_integration_examples.cxx

* do not run tests for servers without collections support

* disable test for capella as it cannot load sample buckets