CXXCBC-562: (Columnar) Provide HTTP session manager with updated cluster (#638)
Motivation
==========
After bootstrap completes, for alternate addressing, cluster options can be
updated.  However, currently for the Columnar connection path those
cluster options are not propagated to the HTTP session manager.  The
lack of propagation means HTTP connections are not able to connect when
using alternate addressing because the external hostname and/or port is
not used.

Changes
=======
* For Columnar connections, pass the updated cluster options back in the
create_sessions() callback so that the options can be propagated to the
HTTP session manager.