CXXCBC-593: apply configuration profile as the last step
To ensure that profile works, we must apply it the options as the very
last step, because the way how CLI parser works, the switches are always
populated, either by user value or by default. And being
--configuration-profile switch itself, we cannot update defaults for
other switches at before the parsing stage, therefore we left with the
only option to defer profile application to the very last step.