Use timestamp from the tag for reproducible tarballs
To improve stability of the tarballs, we need to make sure that that the
package does not have changing parts for releases. But at the same time,
it is useful to have build date recorded. This patch instructs cmake to
use timestamp from the tag, if it exists (GA packages), so the checksum
of the tarballs would never change even if the tarball will be produced
later on different machine. Embedded certificates would still affect
reproducibility, but it is unlikely that they would change frequently.